requirements.txt
lxml==4.2.1
numpy==1.14.3
pandas==0.23.0
python-dateutil==2.7.3
python-docx==0.8.6
pytz==2018.4
six==1.11.0
xlrd==1.1.0
